Add option for setting pre trigger record duration
The new option can be accessed giving `-p <MILLIS>` as an input
argument. It sets the maximum duration for which packets received
before the trigger will be displayed in the hexdump.
However, it is not guaranteed that the hexdump will contain packets
for the set amount of time before the trigger since the trigger
might already have occured shortly after starting the recording.